+++ domain/src/main/java/daily/dayo/domain/usecase/folder/RequestFolderInfoUseCase.kt
@@ -6,6 +6,6 @@ import javax.inject.Inject
 class RequestFolderInfoUseCase @Inject constructor(
     private val folderRepository: FolderRepository
 ) {
-    suspend operator fun invoke(folderId: Int) =
+    suspend operator fun invoke(folderId: Long) =
         folderRepository.requestFolderInfo(folderId)
 }
